requirements.txt
openai
faiss-cpu
numpy
streamlit
